Add a simple sign up action
This commit adds a small sign up action as well as a custom form builder which can be utilised to display errors in a bulma form style. It also customises the rails field_error_proc in order to give the correct error class to the field with errors instead of rails' default wrapping in a div approach.